GUI changes, and updated chicobot
Added more to the gui, I figured out the tkinter radio buttons so that
the values are actually usable. I also added the checkbox for full object
data as well as the browse button that works! Finally, for chicobot, I
changed the weekly message, picture, and added a greeting that is set
to the time of day, so at 6:45 am it says Good Morning, at 4:00 pm it says
Good Evening... ect.