Big changes! (Starting Gui)
Gui.py is just the gui seperated from the whole program, that's where I
actually built it. This was a pretty long project, I think I started a
few weeks ago (on 09-23). I was able to put the option for the whole df
but I haven't yet implemented that into the real program. I'm very happy
with how the gui turned out, although it was so many lines for something
so simple.

The plan is to work on outputing the filament length for each of the
objects, I predict that will be in a month or so (unless I get lots of
motivation), since I have exams coming up in the next few weeks.

-Ryan (10/07/2022)