Adding Object Size to Output Excel
This was a bit of a project, but I finally was able to output the size
of each tracked object in pixels. Since the objects tracked over many
frames, I output the object size mean and std. I was having problems by
not excluding the objects which were only tracked for one frame, but I
figured it out eventually.

Also there are a few changes to chicobot, mainly just typos for this
weeks email!

-Ryan (10/15/2022)