Error when saving >1 sheet:
I'm getting an error when I try to concatenate the df's. It works on the
first loop, but after that, Im getting the error "Exception has occurred:
InvalidIndexErrorReindexing only valid with uniquely valued Index objects".

I'm guessing that there are leftover values that exist and are not being
reset after the first loop. I'm also thinking I need to change the way
that I save dfs to multiple excel sheets, since I've always gotten errors
when opening the saved values. (Maybe using a different dataframe thats
generated by the for loop, and then having another for loop to save them)