Adding File Folder Creation!
It was actually super easy to add the folder creation, im kind of mad
that i didnt do it sooner! Anyways, it works well now, and I added a field
to the starting GUI so you can customize the folder name. It defaults to
today's date tho, incase you dont want to mess with it.