Changing MSD Lines
We realized that we dont want mean squared displacement, instead we just
want the regular displacement of each object. I have started changing it
so that the program just uses the pythagorean theorem to calculate the
change in distance from frame to frame. I think that I've gotten the
calculations right but I still need to work on formatting the data to
output.