Designed New DataFrame
I finished outputting the new displacement data as a DataFrame, the only
thing I have left to figure out is how to rename the top indices. I want
to have it automatically done, so that I dont need to 1: type out 50
keys for the index names and 2: so that if the number of frames present
or fps changes, the program won't die.

Anyways thats pretty small, but the dataframe is finished!