Fixing the problem with dropped frames!
Just adding the difference between frames and dividing the instant speed
by that number. I also added more comments where the linking and tracking
of the particles. However, I do need to find a solution to keeping the
index between the full filament data and the regular data the same. This
started because I need to reset the index now that I changed from MSD to
instantaneous velocities.