Adding Distance Column
I added a distance column that just tracks the change in position from
the first frame to the last frame. Next up on the to-do list is to create
a separate Phil_main that will be able to run multiple days worth of
data in a row. Besides that, I also need to allow different video formats
and maybe display the thresholded frames as they are being tracked.