Finished Naming Convention/Documentation (wow) & Bug Fix
I added kind of a long explanation for the naming convention function,
but I figure it would be very confusing to figure out without an
explanation. Also I fixed a bug where the compiled program would have a
NameError, because I was using exit() but it should've been sys.exit().