Okay, this time I actually fixed the bug!
I was able to add some error handling, where the TypeError was being
caused by returning book as nothing. I also fixed the delete command,
and changed it from .pop to just using the del keyword. Idk why the
pop was triggering the TypeError, 'builtin_function_or_method' object is
not subscriptable, but as long as it's fixed, I'm happy lol