Adding filetype memory, changed settings datatype
Phil now remembers if you used AVI files in previous runs, and the
settings are now saved as formatted json files. Inside Phil, settings is
now saved as a dict rather than a list, which I think will be easier to
work with in the future, if we (I) decide to keep making changes.

On a side note, I noticed that folders are now saved in my VScode dir,
rather than in Phil modules. I'm not sure if this is a product of my
changes, but I'll investigate later.