Making plot_traj automated, all UIs now screen adaptive
The adaptive sizing is something I've been itching to change for a while
and it now works pretty well with higher resolution screens. Previously
the tkinter UIs were blurry on 4k and high res screens, but with the
ctypes library, I was able to change it.

For the automated plot_traj file creation, i had to read through the
trackpy source code to finally understand how it works. Now need to
add option for saving these files, and perhaps a DPI adjustment.